QUALITY: Added Guards + Made Cohesive Coding Standards
