+++ eagle/eagleAPI/src/main/java/eagle/sdkInterface/sensorAdaptors/AdaptorRPLIDAR.java
@@ -32,10 +32,6 @@ public abstract class AdaptorRPLIDAR extends SensorAdaptor {
         }
     }
 
-    public boolean setAndroidContext(Object object) {
-        return false;
-    }
-
     public float[] getCalibrationOffset() {
         return calibrationOffset;
     }
